requirements.txt
discord.py
python-dotenv
supabase
supabase_auth
supabase_functions
asyncpg
python-dateutil